Split home.nix into linux and darwin version
I tried for a bit doing a single `home.nix` file and using `isLinux` /
`isDarwin` to toggle certain packages on certain platforms only, but
couldn't get it to work. There's some sort of fixed point calculation
going on that the `home.nix` file is part of that I don't fully
understand, but seems responsible for weird errors.

`mkMerge` and `mkIf` are the functions to learn more about if I want to
take another go at some point.

That said, the split file solution might even be the nicer solution.