Use overlays to implement custom package changes
This is nice. We can define how each package should be modified in
isolation, and Nix will take care of their interdependencies.

The `overlays` directory needs to be symlinked to
`~/.config/nixpkgs/overlays`. Looks like there might be ways to define
it as part of the nix configuration using the nixos configuration.