Improve configuration management in Nix
The `--config` parameter only seems to do a thing on the initial setup
of the filebrowser package, if no database yet exists. If a filebrowser
database is already created the option is ignored. As a result
configuration changes in Nix won't be automatically propagated into the
filebrowser package, which isn't very nixy.

Luckily the filebrowser package has an `import` command that when passed
a configuration file will use it to completely replace existing
configuration options. Unfortunately the format of the file isn't really
documented anywhere, so requiring users to pass in such a configuration
file would likely be a source of frustration.

We can ask the users to provide configuration using the options that are
found in the filebrowser documentation. We can apply these on top of a
fresh filebrowser configuration, then export the modified configuration
back into a file.