Replace minimserver with navidrome for music
Problem with a DLNA-based approach like minimserver is that it was never
going to work with Tailscale, at least not current versions. I thought
for a while the new UPnP support in Tailscale would make this possible,
but didn't realize that's a different sort of UPnP functionality. For
DLNA music discovery you need Multicast, which is something Tailscale
does not at present support.