Sandbox yarr rss reader application
Today I learned about the `systemd-analyze security` command, and using
it as a guide I limited access of the yarr service as much as possible.
It still needs to be able to talk to the network, to be accessible and
load RSS feeds. Other than that, it can't do much anymore.

I should take a look at my other services too, but that's for some other
time.