Fix vale configuration
I remember Vale working at some point in the past, but maybe due to
updates in my configuration, maybe due to updates with Vale, or both, it
hasn't been working for a while.

This enables vale style checking on markdown files and git commit
messages. It enables a bunch of different style checkers too.