sentient-tshirt: don't lock before sleeping
I've got the laptop setup now to hibernate when I close the lid.
Resuming from that requires unlocking the disk again, because the
hibernation image is stored in encrypted swap. Same as when I start the
machine I don't want to prompt for a password _again_ after unlocking
the disk.