qutebrowser: fix tab-move hotkeys
When a number is passed in addition to + and -, it moves tabs to to the
position indicated by that number. +1 means making a tab the first tab,
-1 making a tab the last tab (- counts from the right).

That's not what I intended, I want to move the current tab to the left
or right with these hotkeys. That's done by only specifying a + or -. I
think I didn't figure this out earlier because I only briefly tested the
situation with two tabs.