ai-banana: set 700 permissions on /persist
This means regular users won't be able to peruse the files in /persist,
and only get access to directories bind-mounted into their private
filesystem.

The best way to do this, I learned, is to give services a RootDir. Then
we can mount a directory /persist/foo into that service without needing
to worry that the /persist directory on the root filesystem is
inaccessible. On the process' filesystem the /persist directory will be
created with other permissions, and only the selected directory mounted
under it.