ai-banana: fix fdm
I accidentally broke this when working on isolating /persist. This sets
up fdm with a private root and access to a couple of things it needs.

It's likely possible to sandbox fdm even better, and I'd like to give
this a short in the future. Particularly there's a couple of security
settings I might add to the unit file (the navidrome unit file on
nixpkgs is a good example). Also it'd be nice to assign it a separate
user, but I don't know what the best pattern is for systemd services
that require shared write-access to a directory.