ai-banana: add append-only modifiers to email dirs
At one point a syncing mishap caused some of my email archive to be
deleted, and I had to recover it from backups. This resulted in me
searching for a way to mark a directory as append-only to prevent it
from happening again. I found that I could set the 'a' file attribute
using the 'chattr' command line application do this. That's what I did,
but it felt unsatisfying not knowing a way to document I had this
property set and why it was important.

Today I learned about tmpfiles.d, which is the way NixOS itself manages
some paths it needs to set up to work declaratively. I'm using it here
to declaratively set the append-only flag on those directories, and
ensure they get created while at it.