ai-banana: Fix calendar generation
It wasn't working because it was still relying on directly accessing a
/persist path, which now only root can do.

It also gets its own user now!

I think it'd be easier/nicer if this were a CGI service. Then the entire
thing wouldn't need access to a filesystem at all and could run as a
DynamicUser (as systemd calls it).