ai-banana: don't set execute bit for files
By setting the mode as a mask, it will only every take permissions away
from existing files, never add them. That way I can set the execute bit
for directories (meaning you can peek into them) without setting it for
files.

The downside of this is that I can't ensure read/write permissions are
set for the syncdata group, which can result in files being created that
won't get synced.

I don't know if this is a good change overall. I'm going to try it for a
bit and see.