sentient-tshirt: try without some kernel params
I don't use sleep mode for this machine, I only hibernate it, so I
don't think I need the sleep-mode related setting. Also I'm getting the
occasional screen-freeze, so whatever the screen-freeze related setting
is supposed to be doing doesn't appear to be working. Let's try with the
defaults for a bit.