syncthing: set umask to 0007
This makes it so synced files will be readable and writable by the user
and group by default. This is great because created files are owned by
the syncthing user and syncdata group, and many other services run as
that same syncdata group in order to be able to read those files.