ai-banana: bind-mount directory paths
We were bind-mounting filepaths, but this caused trouble when syncthing
would write to a bind-mounted filepath. Syncthing replaces the file, and
so the service with the bind mount ends up with a different copy from
the one now synced by syncthing. This fixes the issue, although the todo
service now has access to more files than really it should. I don't care
too strongly at the moment because the todo service is running my own
code, but it's something to consider for the future.