neovim: Create session and exit on suspend
I'm using nushell, which does not support background processes (yet).
When I suspend Vim (which is in my muscle memory), I get a blank
terminal and have to Ctrl+C my way out of it.

This makes it so neovim will dump a session and exit when I try to
suspend it using Ctrl+Z. I can get back in by starting it with `neovim
-S`.