airborne-cactus: Document 802.11ac research
I've currently configured the wifi access point to use 802.11g. It
works, and it's good enough for now because my most-used machines are on
ethernet, but it'd be nice to make use of the 802.11ac support of my
wireless card at some point for faster Wifi.

Turns out this is not so easy to set up. This commit includes some links
to useful resources for giving it another try another time.