airborne-cactus: don't wait for tailscale0
When running nixos switch on airborne-cactus it currently hangs on
waiting for the tailscale interface to come online. systemd-networkd
doesn't manage or exactly seem to know what it means for tailscale to be
online.

I want the online-check to only check the wan interface. The lan and
wlan interfaces can remain offline if there's no cable connected, so I
ignore them from the networkd config itself (tailscale0 does not have an
entry there).

Allowlisting the interfaces to check doesn't appear to be possible.
Setting systemd-network.wait-online.anyInterface comes closest, but is a
bit to undirected for my taste for a router.