airborne-cactus: Fix startup race condition
networkd was responsible for adding wlan to the bridge, but this only
worked after hostapd had started which in practice would be later. This
meant that bridge-connected wifi would not set itself up on boot, and
I'd have to ssh in over a wired connection to restart systemd-networkd
to retry hooking wlan up to the bridge.

This takes an alternative approach, of letting hostapd add the interface
to the bridge when it's ready, resolving the timing issue.

Not a hundred percent sure what would happen if hostapd would try to add
to the bridge before networkd is ready for it, or if that's even
possible. I think/hope hostapd might crash and be restarted by systemd,
which would effectively cause it to retry again a bit later. The restart
behavior seems to be in place, looking at the nixos module, so the
crashing is the open question.