sentient-tshirt: randomize MAC
There's an option that creates a persistent but different MAC address
for each network, but the documentation mentions a downside of this
approach is the MAC change takes about 300ms which adds to connection
time and can create race conditions.

This version of the behavior sets a new MAC every time iwd starts, which
seems nice. Also maybe even closer to the behavior I want.