passage: install
I've always liked using `pass`, but got frustrated maintaining GPG keys
and eventually threw it out. I've been using keepassxc since with a
small custom wrapper for a slightly pass-like experience. This was
okay, but I've liked using pass better. I've been keeping an eye on the
age ecosystem, and it's providing a great pass alternative not based of
gpg, so I'm going to give that a try.