[WIP] gonic: set up gonic-specific Tailscale
This doesn't quite work yet. I need to set up container networking. This
is an area where nixos is making changes though, looking to let
container networking make use of networkd, so I don't know how much I
want to push here, things might be unstable.

Another edge is tailscale requiring oauth in order to get auth keys for
a container each time it starts. There's a get-authkey tool for this,
but it isn't packed up and available for nixos yet. You can `go run` it,
but that might be inefficient.