nu: remove formats plugin
I like the plugin, but registering it in config.nu does not work
reliably and will sometimes result in nushell crashing.

Nushell doesn't currently appear to support a way of defining plugins
declaratively. There's some workarounds below, but they're too
complicated for my taste:
https://github.com/nushell/nushell/discussions/12997

I'm going to wait this out and see if there's a better way to do this in
a  couple months' time.