aerc: Mark archived email as seen
When marking an email as seen, an S gets added to the end of its
filename. Due to syncing mixups it's possible for a seen and unseen
version of an email to exist in the same directory, which causes the
email to show up in a weird way in aerc. By alwaysing marking email as
seen when archiving we can avoid this.