jujutsu: use default pager setting
jujutsu uses a default pager setting 'less -FRX'. The -F option is particulary
useful, it ensures output is simply printed to terminal if it fits a single
screen. I was missing that because NixOS sets $PAGER to 'less' by default,
which overrules the default setting. Adding an explicit configuration value
overrules $PAGER in turn.